+++ xray-detector/display.h
@@ -15,11 +15,11 @@ void display_on();
 void display_off();
 bool display_is_on();
 
-void display_fill_rect(uint8_t x, uint8_t y, uint8_t w, uint8_t h, uint16_t color);
-void display_draw_rect(uint8_t x, uint8_t y, uint8_t w, uint8_t h, uint16_t color);
-void display_draw_bitmap(uint8_t x, uint8_t y, const uint8_t bitmap[], uint8_t w, uint8_t h, uint16_t color);
-void display_draw_line(uint8_t x0, uint8_t y0, uint8_t x1, uint8_t y1, uint16_t color);
-void display_set_cursor(uint8_t x, uint8_t y);
+void display_fill_rect(uint16_t x, uint16_t y, uint16_t w, uint16_t h, uint16_t color);
+void display_draw_rect(uint16_t x, uint16_t y, uint16_t w, uint16_t h, uint16_t color);
+void display_draw_bitmap(uint16_t x, uint16_t y, const uint8_t bitmap[], uint8_t w, uint8_t h, uint16_t color);
+void display_draw_line(uint16_t x0, uint16_t y0, uint16_t x1, uint16_t y1, uint16_t color);
+void display_set_cursor(uint16_t x, uint16_t y);
 void display_set_textcolor(uint16_t color);
 void display_set_textsize(uint8_t size);
 void display_prints(const char * text);
@@ -27,7 +27,7 @@ void display_println(const char * text);
 void display_print8(uint8_t value);
 void display_print16(uint16_t value);
 void display_print32(uint32_t value);
-uint8_t display_get_cursor_x();
-uint8_t display_get_cursor_y();
+uint16_t display_get_cursor_x();
+uint16_t display_get_cursor_y();
 
 #endif /* DISPLAY_H_ */
